Vererbung bei Figure Tests eingebaut - Phase 1
