+++ app/src/test/java/com/uni/gruppenphaseandroid/playingfieldtests/FigureTestMovingPossible.java
@@ -1,0 +1,465 @@
+package com.uni.gruppenphaseandroid.playingfieldtests;
+// TODO: Optionen für Startfeld und Zielfelder testen => grüne Figuren
+
+import com.uni.gruppenphaseandroid.cards.Card;
+import com.uni.gruppenphaseandroid.cards.Cardtype;
+import com.uni.gruppenphaseandroid.manager.GameManager;
+import com.uni.gruppenphaseandroid.playingfield.Citizen;
+import com.uni.gruppenphaseandroid.playingfield.Color;
+import com.uni.gruppenphaseandroid.playingfield.Jerk;
+import com.uni.gruppenphaseandroid.playingfield.King;
+import com.uni.gruppenphaseandroid.playingfield.Knight;
+import com.uni.gruppenphaseandroid.playingfield.Typ;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+public class FigureTestMovingPossible extends FigureTest {
+
+    @Before
+    public void setUp(){ // TODO: NOK
+        Card greencard = new Card(Cardtype.ONEORELEVEN_START);
+        GameManager.getInstance().setSelectedCard(greencard);
+    }
+
+    /**
+     * Test innerhalb einer Farbe auf Move
+     * 2 => Überholen; 1 => Rausschmeißen
+     */
+    @Test
+    public void checkMovingRedJerkByRedCitizen2() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        citizenRed = new Citizen(2, Color.RED, field1, Typ.CITIZEN, figureUI2);
+        Assert.assertTrue(citizenRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedJerkByRedCitizen1() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        citizenRed = new Citizen(2, Color.RED, field1, Typ.CITIZEN, figureUI2);
+        Assert.assertTrue(citizenRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedJerkByRedKnight2() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        knightRed = new Knight(3, Color.RED, field1, Typ.KNIGHT, figureUI3);
+        Assert.assertTrue(knightRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedJerkByRedKnight1() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        knightRed = new Knight(3, Color.RED, field1, Typ.KNIGHT, figureUI3);
+        Assert.assertTrue(knightRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedJerkByRedKing2() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        kingRed = new King(4, Color.RED, field1, Typ.KING, figureUI4);
+        Assert.assertTrue(kingRed.checkMoving(2));
+    }
+    @Test
+    public void checkMovingRedJerkByRedKing1() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        kingRed = new King(4, Color.RED, field1, Typ.KING, figureUI4);
+        Assert.assertTrue(kingRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedCitizenByRedJerk2() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        jerkRed = new Jerk(1, Color.RED, field1, Typ.JERK, figureUI1);
+        Assert.assertFalse(jerkRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedCitizenByRedJerk1() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        jerkRed = new Jerk(1, Color.RED, field1, Typ.JERK, figureUI1);
+        Assert.assertTrue(jerkRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedCitizenByRedKnight2() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        knightRed = new Knight(3, Color.RED, field1, Typ.KNIGHT, figureUI3);
+        Assert.assertTrue(knightRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedCitizenByRedKnight1() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        knightRed = new Knight(3, Color.RED, field1, Typ.KNIGHT, figureUI3);
+        Assert.assertTrue(knightRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedCitizenByRedKing2() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        kingRed = new King(4, Color.RED, field1, Typ.KING, figureUI4);
+        Assert.assertTrue(kingRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedCitizenByRedKing1() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        kingRed = new King(4, Color.RED, field1, Typ.KING, figureUI4);
+        Assert.assertTrue(kingRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKnightByRedJerk2() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        jerkRed = new Jerk(1, Color.RED, field1, Typ.JERK, figureUI1);
+        Assert.assertFalse(jerkRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKnightByRedJerk1() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        jerkRed = new Jerk(1, Color.RED, field1, Typ.JERK, figureUI1);
+        Assert.assertTrue(jerkRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKnightByRedCitizen2() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        citizenRed = new Citizen(2, Color.RED, field1, Typ.CITIZEN, figureUI2);
+        Assert.assertFalse(citizenRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKnightByRedCitizen1() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        citizenRed = new Citizen(2, Color.RED, field1, Typ.CITIZEN, figureUI2);
+        Assert.assertTrue(citizenRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKnightByRedKing2() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        kingRed = new King(4, Color.RED, field1, Typ.KING, figureUI4);
+        Assert.assertTrue(kingRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKnightByRedKing1() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        kingRed = new King(4, Color.RED, field1, Typ.KING, figureUI4);
+        Assert.assertTrue(kingRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKingByRedJerk2() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        jerkRed = new Jerk(1, Color.RED, field1, Typ.JERK, figureUI1);
+        Assert.assertFalse(jerkRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKingByRedJerk1() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        jerkRed = new Jerk(1, Color.RED, field1, Typ.JERK, figureUI1);
+        Assert.assertFalse(jerkRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKingByRedCitizen2() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        citizenRed = new Citizen(2, Color.RED, field1, Typ.CITIZEN, figureUI2);
+        Assert.assertFalse(citizenRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKingByRedCitizen1() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        citizenRed = new Citizen(2, Color.RED, field1, Typ.CITIZEN, figureUI2);
+        Assert.assertFalse(citizenRed.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKingByRedKnight2() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        knightRed = new Knight(3, Color.RED, field1, Typ.KNIGHT, figureUI3);
+        Assert.assertFalse(knightRed.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKingByRedKnight1() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        knightRed = new Knight(3, Color.RED, field1, Typ.KNIGHT, figureUI3);
+        Assert.assertFalse(knightRed.checkMoving(1));
+    }
+
+
+    /**
+     * Test Rausschmeißen zwischen Farben
+     * Ausnahme King kann nur von King geschmissen werden.
+     */
+    @Test
+    public void checkMovingRedJerkByBlueJerk2() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        jerkBlue = new Jerk(5, Color.BLUE, field1, Typ.JERK, figureUI5);
+        Assert.assertTrue(jerkBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedJerkByBlueJerk1() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        jerkBlue = new Jerk(5, Color.BLUE, field1, Typ.JERK, figureUI5);
+        Assert.assertTrue(jerkBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedJerkByBlueCitizen2() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        citizenBlue = new Citizen(6, Color.BLUE, field1, Typ.CITIZEN, figureUI6);
+        Assert.assertTrue(citizenBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedJerkByBlueCitizen1() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        citizenBlue = new Citizen(6, Color.BLUE, field1, Typ.CITIZEN, figureUI6);
+        Assert.assertTrue(citizenBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedJerkByBlueKnight2() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        knightBlue = new Knight(7, Color.BLUE, field1, Typ.KNIGHT, figureUI7);
+        Assert.assertTrue(knightBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedJerkByBlueKnight1() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        knightBlue = new Knight(7, Color.BLUE, field1, Typ.KNIGHT, figureUI7);
+        Assert.assertTrue(knightBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedJerkByBlueKing2() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        kingBlue = new King(8, Color.BLUE, field1, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedJerkByBlueKing1() { // OK
+        jerkRed = new Jerk(1, Color.RED, field2, Typ.JERK, figureUI1);
+        kingBlue = new King(8, Color.BLUE, field1, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedCitizenByBlueJerk2() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        jerkBlue = new Jerk(5, Color.BLUE, field1, Typ.JERK, figureUI5);
+        Assert.assertFalse(jerkBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedCitizenByBlueJerk1() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        jerkBlue = new Jerk(5, Color.BLUE, field1, Typ.JERK, figureUI5);
+        Assert.assertTrue(jerkBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedCitizenByBlueCitizen2() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        citizenBlue = new Citizen(6, Color.BLUE, field1, Typ.CITIZEN, figureUI6);
+        Assert.assertTrue(citizenBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedCitizenByBlueCitizen1() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        citizenBlue = new Citizen(6, Color.BLUE, field1, Typ.CITIZEN, figureUI6);
+        Assert.assertTrue(citizenBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedCitizenByBlueKnight2() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        knightBlue = new Knight(7, Color.BLUE, field1, Typ.KNIGHT, figureUI7);
+        Assert.assertTrue(knightBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedCitizenByBlueKnight1() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        knightBlue = new Knight(7, Color.BLUE, field1, Typ.KNIGHT, figureUI7);
+        Assert.assertTrue(knightBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedCitizenByBlueKing2() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        kingBlue = new King(8, Color.BLUE, field1, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedCitizenByBlueKing1() { // OK
+        citizenRed = new Citizen(2, Color.RED, field2, Typ.CITIZEN, figureUI2);
+        kingBlue = new King(8, Color.BLUE, field1, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKnightByBlueJerk2() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        jerkBlue = new Jerk(5, Color.BLUE, field1, Typ.JERK, figureUI5);
+        Assert.assertFalse(jerkBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKnightByBlueJerk1() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        jerkBlue = new Jerk(5, Color.BLUE, field1, Typ.JERK, figureUI5);
+        Assert.assertTrue(jerkBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKnightByBlueCitizen2() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        citizenBlue = new Citizen(6, Color.BLUE, field1, Typ.CITIZEN, figureUI6);
+        Assert.assertFalse(citizenBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKnightByBlueCitizen1() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        citizenBlue = new Citizen(6, Color.BLUE, field1, Typ.CITIZEN, figureUI6);
+        Assert.assertTrue(citizenBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKnightByBlueKnight2() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        knightBlue = new Knight(7, Color.BLUE, field1, Typ.KNIGHT, figureUI7);
+        Assert.assertTrue(knightBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKnightByBlueKnight1() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        knightBlue = new Knight(7, Color.BLUE, field1, Typ.KNIGHT, figureUI7);
+        Assert.assertTrue(knightBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKnightByBlueKing2() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        kingBlue = new King(8, Color.BLUE, field1, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKnightByBlueKing1() { // OK
+        knightRed = new Knight(3, Color.RED, field2, Typ.KNIGHT, figureUI3);
+        kingBlue = new King(8, Color.BLUE, field1, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKingByBlueJerk2() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        jerkBlue = new Jerk(5, Color.BLUE, field1, Typ.JERK, figureUI5);
+        Assert.assertFalse(jerkBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKingByBlueJerk1() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        jerkBlue = new Jerk(5, Color.BLUE, field1, Typ.JERK, figureUI5);
+        Assert.assertFalse(jerkBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKingByBlueCitizen2() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        citizenBlue = new Citizen(6, Color.BLUE, field1, Typ.CITIZEN, figureUI6);
+        Assert.assertFalse(citizenBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKingByBlueCitizen1() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        citizenBlue = new Citizen(6, Color.BLUE, field1, Typ.CITIZEN, figureUI6);
+        Assert.assertFalse(citizenBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKingByBlueKnight2() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        knightBlue = new Knight(7, Color.BLUE, field1, Typ.KNIGHT, figureUI7);
+        Assert.assertFalse(knightBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKingByBlueKnight1() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        knightBlue = new Knight(7, Color.BLUE, field1, Typ.KNIGHT, figureUI7);
+        Assert.assertFalse(knightBlue.checkMoving(1));
+    }
+
+
+    @Test
+    public void checkMovingRedKingByBlueKing2() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        kingBlue = new King(8, Color.BLUE, field1, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingRedKingByBlueKing1() { // OK
+        kingRed = new King(4, Color.RED, field2, Typ.KING, figureUI4);
+        kingBlue = new King(8, Color.BLUE, field1, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(1));
+    }
+
+    @Test
+    public void checkMovingGreenKingByBlueKing13() { // NOK
+        kingGreen = new King(12, Color.GREEN, field13, Typ.KING, figureUI12);
+        kingBlue = new King(8, Color.BLUE, field12, Typ.KING, figureUI8);
+        Assert.assertFalse(kingBlue.checkMoving(2));
+    }
+
+    @Test
+    public void checkMovingGreenKingByBlueKing14() { // NOK
+        kingGreen = new King(12, Color.GREEN, field14, Typ.KING, figureUI12);
+        kingBlue = new King(8, Color.BLUE, field13, Typ.KING, figureUI8);
+        Assert.assertTrue(kingBlue.checkMoving(2));
+    }
+}
